Let rtlsdr_set_tuner_if_gain return -1 upon error.
Let rtlsdr_set_tuner_if_gain() return -1 when it cannot be set.
That was the case when there is no device, no tuner or when the
value was invalid, but it returned 0 when the there was no way
to set if-gain in the receiver.
Now it returns also -1 when there is no set-function for if_gain.